mavlink: move ALTITUDE to streams header
